requirements.txt
pandas
numpy
matplotlib
seaborn
jinja2
flask
plotly
pyarrow
pdfkit
weasyprint
openai
python-dotenv
markdown
gpt4free